+++ statconvert/compare/__init__.py
@@ -10,7 +10,9 @@ from statconvert.compare.comparison import (
 from statconvert.compare.exceptions import CompareError
 from statconvert.compare.models import (
     ColumnComparison,
+    CompareOptions,
     CompareIssue,
+    DifferenceDetail,
     DatasetComparison,
     MetadataComparison,
     SchemaComparison,
@@ -18,14 +20,17 @@ from statconvert.compare.models import (
     ValueComparison,
 )
 from statconvert.compare.reporting import (
+    comparison_to_json_payload,
     infer_compare_report_format,
     write_compare_report,
 )
 
 __all__ = [
     "ColumnComparison",
+    "CompareOptions",
     "CompareError",
     "CompareIssue",
+    "DifferenceDetail",
     "DatasetComparison",
     "MetadataComparison",
     "SchemaComparison",
@@ -37,6 +42,7 @@ __all__ = [
     "compare_schema",
     "compare_shape",
     "compare_values_summary",
+    "comparison_to_json_payload",
     "resolve_compare_object_selectors",
     "infer_compare_report_format",
     "write_compare_report",